ancora un refactoring provato su Debian
