+++ coa/docs/md/README.md
@@ -2,15 +2,6 @@
 
 Benvenuto nella guida ai comandi di **coa**. Qui trovi l'elenco completo delle funzionalità orchestrate dal motore.
 
-## ⚙️ System Tools
-
-- [**adapt**](coa_adapt.md) - Adapt the video resolution to the Virtual Machine window
-- [**build**](coa_build.md) - Compile binaries and generate native distribution packages (.deb, PKGBUILD)
-- [**detect**](coa_detect.md) - Display detected host system information
-- [**help**](coa_help.md) - Help about any command
-- [**kill**](coa_kill.md) - Free the nest and unmount filesystems
-- [**version**](coa_version.md) - Print the version number of coa
-
 ## 🚀 Core Actions
 
 - [**remaster**](coa_remaster.md) - Start a system remastering flight (ISO production)
@@ -19,4 +10,13 @@ Benvenuto nella guida ai comandi di **coa**. Qui trovi l'elenco completo delle f
 ## 📦 Export & Sync
 
 - [**export**](coa_export.md) - Export artifacts (iso, pkg) to a remote Proxmox storage
+
+## ⚙️ System Tools
+
+- [**adapt**](coa_adapt.md) - Adapt the video resolution to the Virtual Machine window
+- [**build**](coa_build.md) - Compile binaries and generate native distribution packages (.deb, PKGBUILD)
+- [**detect**](coa_detect.md) - Display detected host system information
+- [**help**](coa_help.md) - Help about any command
+- [**kill**](coa_kill.md) - Free the nest and unmount filesystems
+- [**version**](coa_version.md) - Print the version number of coa
