+++ coa/src/internal/engine/handle_detect.go
